requirements.txt
requests==2.28.2
pyopenssl==23.2.0